# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ElorAdmin API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Shows roles",
                "operationId": "6cd48af64dafcf1c85923dba47e841a4",
                "responses": {
                    "200": {
                        "description": "Shows all roles."
                    },
                    "204": {
                        "description": "No roles found."
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a role",
                "operationId": "4eb1ce389f63b4dbfcc5b002014636e1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Role Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Role created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Shows one role",
                "operationId": "d932a477623c1c931444dcc799f1eb0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request accepted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "That Role doesn't exist."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Edit a role",
                "operationId": "e0fa8b9f9f9cfb6c7e51a7609dd35dcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the Role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Role updated."
                    },
                    "400": {
                        "description": "Role doesnt exist."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete roles",
                "operationId": "a0c87b804be9d37055fc6f936cbacb1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Role deleted."
                    },
                    "403": {
                        "description": "That role can't be deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}